chore: move Dockerfile.cli overlay to root on sync
The RHTAP system only recognizes Dockerfiles in certain locations. This
file needs to be moved to the root directory on sync so that it will be
recognized during builds by RHTAP.

The overlay itself has been moved up to `redhat/overlays` so that if we
need to add additional, we simply copy recursively the entire directory.

Signed-off-by: Lance Ball <lball@redhat.com>