feat: add Dockerfile for backfill-redis
This commit adds a Dockerfile for building a `backfill-redis` image,
which should be used periodically to ensure that redis contains the
complete rekor log.

The patchfile is required due to this upstream issue:

https://github.com/sigstore/rekor/pull/1685

Once that pull request has been merged, we'll need to remove the
patchfile from `main` but not from the `midstream-v1.2.2` branch unless
it is backported upstream.

Signed-off-by: Lance Ball <lball@redhat.com>